services/web/requirements.txt
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
setuptools==61.2.0
Werkzeug==2.1.2
wheel==0.37.1
wincertstore==0.2
Flask-SQLAlchemy==2.5.1
py-linq==1.3.0
dnspython==2.2.1
email-validator==1.2.1
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.2
idna==3.3
mongoengine==0.24.1
numpy==1.23.1
pandas==1.4.3
pip==21.2.4
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
wincertstore==0.2
WTForms==3.0.1
